requirements.txt
--extra-index-url https://test.pypi.org/simple/
allure-pytest==2.16.0
allure-python-commons==2.16.0
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
execnet==2.1.2
greenlet==3.4.0
idna==3.13
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
narwhals==2.20.0
packaging==26.2
playwright==1.58.0
plotly==6.7.0
pluggy==1.6.0
psutil==7.2.2
pyee==13.0.1
Pygments==2.20.0
pytest==8.4.2
pytest-base-url==2.1.0
pytest-html==4.2.0
pytest-metadata==3.1.1
pytest-playwright==0.7.2
pytest-pulse-report==1.0.4
pytest-xdist==3.8.0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.33.1
setuptools==69.5.1
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.6.3